services/analytics/pyproject.toml
[project]
name = "oura-analytics"
version = "0.1.0"
description = "Oura personal analytics service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "psycopg[binary]>=3.1.0",
    "pydantic-settings>=2.1.0",
    "scikit-learn>=1.4.0",
    "statsmodels>=0.14.0",
    "ruptures>=1.1.9",
    "httpx>=0.26.0",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.1.0",
    "cryptography>=42.0.0",
    "openai>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
